pyproject.toml
[project]
name = "docs-chatter"
version = "0.1.0"
description = "Based on RAG Chatbot for Slack"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "atlassian-python-api>=4.0.7",
    "beautifulsoup4>=4.14.3",
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "markdownify>=1.1.0",
    "opensearch-py>=3.1.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "slack-bolt>=1.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
docs-chatter = "docs_chatter.slack.bot:main"
run-batch = "scripts.run_batch:main"

[tool.hatch.build.targets.wheel]
packages = ["src/docs_chatter"]